Add forward statement to for entities to Metamorph
This is a EXPERIMENTAL implementation for passing data structures
through Metamorph. It is very likely to change in the future
substantially. Do not use it in production code!

This commit demonstrates one possible method for enabling Metamorph to
pass complex data structures through a transformation. It does not rely
on the existing named-value pipeline framework for passing literals and
end-of-entity events but introduces a complementary framework for
building pipelines for entities. This pipelines are meant to be used
together with the named-value pipelines for literals.

To demonstrate the possibilities of complex structure handling in
Metamorph a new statement it is added to Metamorph which pipes a complex
data structure through a transformation. The outermost entity of this
complex structure can be renamed while passing it or optionally be
removed, which essentially unwraps the entities contents. See the
test-cases in org/culturegraph/mf/morph/EntityDataTest.xml for examples.